requirements.txt
google-genai
python-dotenv
pandas
fsspec
gcsfs
